+++ packages/vvisp/commands/index.js
@@ -2,7 +2,7 @@ const analyze = require('./analyze');
 const ci = require('./ci');
 const compile = require('./compile');
 const console = require('./console');
-const coverage = require('./coverage')
+const test = require('./test')
 const debug = require('./debug');
 const deployContract = require('./deploy-contract');
 const deployService = require('./deploy-service');
@@ -17,12 +17,12 @@ module.exports = [
   ci,
   compile,
   console,
-  coverage,
   debug,
   deployContract,
   deployService,
   flatten,
   genScript,
   init,
-  showState
+  showState,
+  test
 ];
