+++ packages/vvisp/scripts/ci.js
@@ -3,7 +3,7 @@ const path = require('path');
 const chalk = require('chalk');
 const { execSync } = require('child_process');
 const { printOrSilent } = require('@haechi-labs/vvisp-utils');
-const test = require('./coverage');
+const test = require('./test');
 const deployService = require('./deploy-service');
 const analyze = require('./analyze');
 const constants = require('../config/Constant');
