+++ packages/vvisp/scripts/index.js
@@ -7,7 +7,7 @@ const flatten = require('./flatten');
 const debug = require('./debug');
 const console = require('./console');
 const showState = require('./show-state');
-const coverage = require('./coverage');
+const test = require('./test');
 const analyze = require('./analyze');
 const ci = require('./ci');
 
@@ -21,7 +21,7 @@ module.exports = {
   debug,
   console,
   showState,
-  coverage,
+  test,
   analyze,
   ci
 };
